requirements.txt
asgiref==3.8.1
Django==5.0.4
django-debug-toolbar==4.3.0
django-filter==24.2
django-seed==0.3.1
djangorestframework==3.15.1
drf-yasg==1.21.7
Faker==24.7.1
inflection==0.5.1
Markdown==3.6
packaging==24.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
setuptools==69.5.1
six==1.16.0
sqlparse==0.4.4
toposort==1.10
tzdata==2024.1
uritemplate==4.1.1
typing_extensions==4.10.0
